chore(mobile): 상수 분리
